clean up stream processor class migration
