+++ packages/streams/package.json
@@ -10,12 +10,11 @@
     "./browser/stream-browser-store": "./src/browser/stream-browser-store.ts",
     "./processor": "./src/processor.ts",
     "./processor-runner": "./src/processor-runner.ts",
-    "./stream-processor-v2": "./src/stream-processor-v2.ts",
-    "./processors/browser-event-feed/v2": "./src/processors/browser-event-feed/v2.ts",
+    "./stream-processor": "./src/stream-processor.ts",
+    "./processors/browser-event-feed/implementation": "./src/processors/browser-event-feed/implementation.ts",
     "./processors/browser-raw-events/implementation": "./src/processors/browser-raw-events/implementation.ts",
-    "./processors/browser-raw-events/v2": "./src/processors/browser-raw-events/v2.ts",
     "./processors/core/contract": "./src/processors/core/contract.ts",
-    "./processors/core/v2": "./src/processors/core/v2.ts",
+    "./processors/core/implementation": "./src/processors/core/implementation.ts",
     "./shared/event": "./src/shared/event.ts",
     "./shared/rpc-target": "./src/shared/rpc-target.ts",
     "./shared/stream-processors": "./src/shared/stream-processors.ts",
