+++ packages/streams/src/browser/processor-state-storage.ts
@@ -1,9 +1,6 @@
 import { createSchemaEnsurer } from "./ensure-schema-once.ts";
 import type { SqlClient } from "./stream-browser-db.ts";
-import type {
-  StreamProcessorSnapshot,
-  StreamProcessorStateStorage,
-} from "../stream-processor-v2.ts";
+import type { StreamProcessorSnapshot, StreamProcessorStateStorage } from "../stream-processor.ts";
 
 const DEFAULT_SUBSCRIPTION_KEY = "";
